apps/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dmitry Kuleshov <vdksystem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
confluent-kafka = "^2.11.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-jose = "^3.5.0"
pydantic = "^2.11.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
